+++ src/vws/exceptions/custom_exceptions.py
@@ -42,3 +42,24 @@ class TargetProcessingTimeout(Exception):
     """
     Exception raised when waiting for a target to be processed times out.
     """
+
+
+class ServerError(Exception):  # pragma: no cover
+    """
+    Exception raised when VWS returns a server error.
+    """
+
+    def __init__(self, response: Response) -> None:
+        """
+        Args:
+            response: The response returned by Vuforia.
+        """
+        super().__init__(response.text)
+        self._response = response
+
+    @property
+    def response(self) -> Response:
+        """
+        The response returned by Vuforia which included this error.
+        """
+        return self._response