+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 Next
 ----
 
+* Return a new error (``vws.custom_exceptions.ServerError``) when the server returns a 5xx status code.
+
 2023.12.27
 ------------